refactored video_box and category_box into one component
