+++ src/components/common/index.js
@@ -1,6 +1,5 @@
-export * from './search_bar';
-export * from './video_search_item';
-export * from './category_box';
-export * from './video_list';
 export * from './dropdown_menu';
-export * from './videos_box';
+export * from './search_bar';
+export * from './video_box';
+export * from './video_list';
+export * from './video_search_item';
